Modernize local validation scripts; replace validate_from_ci.py with integrated logic in validate_local.py and update command outputs for clarity
